+++ src/Services/Base/Checker.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Sensorario\Yaphlo\Services\Base;
+
+use Sensorario\Yaphlo\Objects\Message;
+use Sensorario\Yaphlo\Config\Config;
+
+interface Checker
+{
+    public function mustChannelBeHidden(Message $message): bool;
+
+    public function setConfig(Config $config): void;
+}