+++ src/Services/RowBuilder/Exceptions/MissingChannelException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Sensorario\Yaphlo\Services\RowBuilder\Exceptions;
+
+class MissingChannelException extends \Exception
+{
+    /** @var string $message */
+    protected $message = 'Oops! Missing channel!';
+}
+
+