+++ src/Services/RowBuilder/Exceptions/MissingLevelException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Sensorario\Yaphlo\Services\RowBuilder\Exceptions;
+
+class MissingLevelException extends \Exception
+{
+    /** @var string $message */
+    protected $message = 'Oops! Missing level!';
+}
+
+