+++ src/Services/RowBuilder/Exceptions/MissingLineException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Sensorario\Yaphlo\Services\RowBuilder\Exceptions;
+
+class MissingLineException extends \Exception
+{
+    /** @var string $message */
+    protected $message = 'Oops! Missing line!';
+}
+
+