+++ tests/int/ArrayConfigTest.php
@@ -1,10 +1,11 @@
 <?php
 
 use Sensorario\Yaphlo\Logger;
-use Sensorario\Yaphlo\Message;
-use Sensorario\Yaphlo\Writer;
-use Sensorario\Yaphlo\ChannelVisibilityChecker;
-use Sensorario\Yaphlo\WriterAdapter;
+use Sensorario\Yaphlo\Objects\Message;
+use Sensorario\Yaphlo\Services\ChannelVisibilityChecker;
+use Sensorario\Yaphlo\Services\RowBuilder\RowBuilder;
+use Sensorario\Yaphlo\Writers\Writer;
+use Sensorario\Yaphlo\Writers\WriterAdapter;
 use Sensorario\Yaphlo\Listeners\Listener;
 use Sensorario\Yaphlo\Config\ArrayConfig;
 
@@ -23,7 +24,9 @@ class ArrayConfigTest extends PHPUnit\Framework\TestCase
         ];
 
         $logger = new Logger(
-            new Message,
+            new Message(
+                new RowBuilder()
+            ),
             new Writer(
                 new ArrayConfig($config),
                 new WriterAdapter(
@@ -56,7 +59,9 @@ class ArrayConfigTest extends PHPUnit\Framework\TestCase
         ];
 
         $logger = new Logger(
-            new Message,
+            new Message(
+                new RowBuilder()
+            ),
             new Writer(
                 new ArrayConfig($config),
                 new WriterAdapter(