+++ tests/unit/WriterTest.php
@@ -2,11 +2,11 @@
 
 namespace Sensorario\Yaphlo\Tests;
 
-use Sensorario\Yaphlo\Writer;
-use Sensorario\Yaphlo\Message;
+use Sensorario\Yaphlo\Objects\Message;
 use Sensorario\Yaphlo\Config\Config;
-use Sensorario\Yaphlo\FileWriterWrapper;
-use Sensorario\Yaphlo\ChannelVisibilityChecker;
+use Sensorario\Yaphlo\Writers\FileWriterWrapper;
+use Sensorario\Yaphlo\Writers\Writer;
+use Sensorario\Yaphlo\Services\ChannelVisibilityChecker;
 
 class WriterTest extends \PHPUnit\Framework\TestCase
 {
@@ -24,17 +24,17 @@ class WriterTest extends \PHPUnit\Framework\TestCase
             ->getMock();
 
         $this->message = $this
-            ->getMockBuilder(\Sensorario\Yaphlo\Message::class)
+            ->getMockBuilder(\Sensorario\Yaphlo\Objects\Message::class)
             ->disableOriginalConstructor()
             ->getMock();
 
         $this->fileWriterWrapper = $this
-            ->getMockBuilder(\Sensorario\Yaphlo\FileWriterWrapper::class)
+            ->getMockBuilder(\Sensorario\Yaphlo\Writers\FileWriterWrapper::class)
             ->disableOriginalConstructor()
             ->getMock();
 
         $this->checker = $this
-            ->getMockBuilder(\Sensorario\Yaphlo\ChannelVisibilityChecker::class)
+            ->getMockBuilder(\Sensorario\Yaphlo\Services\ChannelVisibilityChecker::class)
             ->disableOriginalConstructor()
             ->getMock();
     }