+++ CHANGELOG.md
@@ -7,3 +7,6 @@
  - [feat] add documentation in readme file
  - [feat] use phpstan for static analysis
  - [feat] documentation in readme file
+ - [feat] throw exception whenever message level is not defined
+ - [feat] add row builder service
+ - [feat] keep objects in specific objects folder